space cleanings and cpp for plugins
