+++ src/thunder/include/plugin_registry.h
@@ -56,7 +56,6 @@ inline const std::map<std::string, std::shared_ptr<BaseGenerator>> GENERATORS = 
 
 
 
-
 /**
  * @brief Helper function to find a loader by name.
  * It searches all three maps and returns the first match.
@@ -70,7 +69,6 @@ inline std::shared_ptr<BaseLoader> find_loader(const std::string& plugin_type) {
 	}
 	return nullptr;
 }
-
 
 /**
  * @brief Helper function to find a builder by name.
